+++ blackjs.js
@@ -1,6 +1,6 @@
 
 // Copyright (c) 2017 Joaquin Giordano
-// version:1.3.1
+// version:1.4.0
 
 
 function checkIdClassOther(text){
@@ -71,6 +71,52 @@ function insert(what,id,eclass,where,data,plusdata) {
 			}
 		}
 		p.innerHTML = data;
+	}
+	if(what == "ul"){
+		var ul = document.createElement(what);
+		if(id != null){
+			ul.setAttribute('id', id);
+		}
+		if(eclass != null){
+			ul.setAttribute('class', eclass);
+		}
+		if(checkWhere == "id"){
+			where = where.substr(1);
+			document.getElementById(where).appendChild(ul);
+		}
+		if(checkWhere == "class"){
+			where = where.substr(1);
+			document.getElementsByClassName(where)[0].appendChild(ul);
+		}
+		if(checkWhere == "other"){
+			if(where == "body"){
+				document.body.appendChild(ul);
+			}
+		}
+	
+	}
+	if(what == "li"){
+		var li = document.createElement(what);
+		if(id != null){
+			li.setAttribute('id', id);
+		}
+		if(eclass != null){
+			li.setAttribute('class', eclass);
+		}
+		if(checkWhere == "id"){
+			where = where.substr(1);
+			document.getElementById(where).appendChild(li);
+		}
+		if(checkWhere == "class"){
+			where = where.substr(1);
+			document.getElementsByClassName(where)[0].appendChild(li);
+		}
+		if(checkWhere == "other"){
+			if(where == "body"){
+				document.body.appendChild(li);
+			}
+		}
+		li.innerHTML = data;
 	}
 	if(what == "button"){
 		var button = document.createElement(what);
@@ -308,9 +354,143 @@ function insert(what,id,eclass,where,data,plusdata) {
 	}
 }
 
+function sha1(msg) {
+  function rotate_left(n,s) {
+    var t4 = ( n<<s ) | (n>>>(32-s));
+    return t4;
+  };
+  function lsb_hex(val) {
+    var str="";
+    var i;
+    var vh;
+    var vl;
+    for( i=0; i<=6; i+=2 ) {
+      vh = (val>>>(i*4+4))&0x0f;
+      vl = (val>>>(i*4))&0x0f;
+      str += vh.toString(16) + vl.toString(16);
+    }
+    return str;
+  };
+  function cvt_hex(val) {
+    var str="";
+    var i;
+    var v;
+    for( i=7; i>=0; i-- ) {
+      v = (val>>>(i*4))&0x0f;
+      str += v.toString(16);
+    }
+    return str;
+  };
+  function Utf8Encode(string) {
+    string = string.replace(/\r\n/g,"\n");
+    var utftext = "";
+    for (var n = 0; n < string.length; n++) {
+      var c = string.charCodeAt(n);
+      if (c < 128) {
+        utftext += String.fromCharCode(c);
+      }
+      else if((c > 127) && (c < 2048)) {
+        utftext += String.fromCharCode((c >> 6) | 192);
+        utftext += String.fromCharCode((c & 63) | 128);
+      }
+      else {
+        utftext += String.fromCharCode((c >> 12) | 224);
+        utftext += String.fromCharCode(((c >> 6) & 63) | 128);
+        utftext += String.fromCharCode((c & 63) | 128);
+      }
+    }
+    return utftext;
+  };
+  var blockstart;
+  var i, j;
+  var W = new Array(80);
+  var H0 = 0x67452301;
+  var H1 = 0xEFCDAB89;
+  var H2 = 0x98BADCFE;
+  var H3 = 0x10325476;
+  var H4 = 0xC3D2E1F0;
+  var A, B, C, D, E;
+  var temp;
+  msg = Utf8Encode(msg);
+  var msg_len = msg.length;
+  var word_array = new Array();
+  for( i=0; i<msg_len-3; i+=4 ) {
+    j = msg.charCodeAt(i)<<24 | msg.charCodeAt(i+1)<<16 |
+    msg.charCodeAt(i+2)<<8 | msg.charCodeAt(i+3);
+    word_array.push( j );
+  }
+  switch( msg_len % 4 ) {
+    case 0:
+      i = 0x080000000;
+    break;
+    case 1:
+      i = msg.charCodeAt(msg_len-1)<<24 | 0x0800000;
+    break;
+    case 2:
+      i = msg.charCodeAt(msg_len-2)<<24 | msg.charCodeAt(msg_len-1)<<16 | 0x08000;
+    break;
+    case 3:
+      i = msg.charCodeAt(msg_len-3)<<24 | msg.charCodeAt(msg_len-2)<<16 | msg.charCodeAt(msg_len-1)<<8  | 0x80;
+    break;
+  }
+  word_array.push( i );
+  while( (word_array.length % 16) != 14 ) word_array.push( 0 );
+  word_array.push( msg_len>>>29 );
+  word_array.push( (msg_len<<3)&0x0ffffffff );
+  for ( blockstart=0; blockstart<word_array.length; blockstart+=16 ) {
+    for( i=0; i<16; i++ ) W[i] = word_array[blockstart+i];
+    for( i=16; i<=79; i++ ) W[i] = rotate_left(W[i-3] ^ W[i-8] ^ W[i-14] ^ W[i-16], 1);
+    A = H0;
+    B = H1;
+    C = H2;
+    D = H3;
+    E = H4;
+    for( i= 0; i<=19; i++ ) {
+      temp = (rotate_left(A,5) + ((B&C) | (~B&D)) + E + W[i] + 0x5A827999) & 0x0ffffffff;
+      E = D;
+      D = C;
+      C = rotate_left(B,30);
+      B = A;
+      A = temp;
+    }
+    for( i=20; i<=39; i++ ) {
+      temp = (rotate_left(A,5) + (B ^ C ^ D) + E + W[i] + 0x6ED9EBA1) & 0x0ffffffff;
+      E = D;
+      D = C;
+      C = rotate_left(B,30);
+      B = A;
+      A = temp;
+    }
+    for( i=40; i<=59; i++ ) {
+      temp = (rotate_left(A,5) + ((B&C) | (B&D) | (C&D)) + E + W[i] + 0x8F1BBCDC) & 0x0ffffffff;
+      E = D;
+      D = C;
+      C = rotate_left(B,30);
+      B = A;
+      A = temp;
+    }
+    for( i=60; i<=79; i++ ) {
+      temp = (rotate_left(A,5) + (B ^ C ^ D) + E + W[i] + 0xCA62C1D6) & 0x0ffffffff;
+      E = D;
+      D = C;
+      C = rotate_left(B,30);
+      B = A;
+      A = temp;
+    }
+    H0 = (H0 + A) & 0x0ffffffff;
+    H1 = (H1 + B) & 0x0ffffffff;
+    H2 = (H2 + C) & 0x0ffffffff;
+    H3 = (H3 + D) & 0x0ffffffff;
+    H4 = (H4 + E) & 0x0ffffffff;
+  }
+  var temp = cvt_hex(H0) + cvt_hex(H1) + cvt_hex(H2) + cvt_hex(H3) + cvt_hex(H4);
+
+  return temp.toLowerCase();
+}
+
+
 function chars(text){
 	var checkText = checkIdClassOther(text);
-	
 	if(checkText == "other"){
 		return text.length;
 	}
@@ -351,7 +531,6 @@ function getElement(idclass){
 
 function contains(where,data){
 	var checkWhere = checkIdClassOther(where);
-	
 	if(checkWhere == "id"){
 		where = where.substr(1);
 		try{
@@ -372,7 +551,6 @@ function contains(where,data){
 	}
 	if(checkWhere == "class"){
 		where = where.substr(1);
-
 			try{
 				var data2 = document.getElementsByClassName(where)[0].value.indexOf(data);
 				if(data2 != undefined){
@@ -383,9 +561,7 @@ function contains(where,data){
 				}
 			}
 			}catch(err){
-
 				var data = document.getElementsByClassName(where)[0].innerHTML.indexOf(data);
-
 				if(data != -1){
 					return true;
 				}else{
@@ -393,9 +569,6 @@ function contains(where,data){
 				}
 			}
 		}
-			
-			
-	
 	if(checkWhere == "other"){
 		var number = where.indexOf(data);
 		if(number != -1){
@@ -435,7 +608,6 @@ var fs = true;
 		}else{
 			actualNumber += e;
 		}
-		
 	})
 	return actualNumber;
 }
@@ -450,7 +622,6 @@ function subs(numbers){
 		}else{
 			actualNumber -= e;
 		}
-		
 	})
 	return actualNumber;
 }
@@ -465,7 +636,6 @@ var fs = true;
 		}else{
 			actualNumber /= e;
 		}
-		
 	})
 	return actualNumber;
 }
@@ -480,7 +650,6 @@ var fs = true;
 		}else{
 			actualNumber *= e;
 		}
-		
 	})
 	return actualNumber;
 }
@@ -597,6 +766,11 @@ function getChar(where,num){
 	return where.charAt(num);
 }
 
+function doTTT(callback,callback2) {
+	callback();
+	callback2();
+}
+
 function time(data){
 
 	var date = new Date();
@@ -652,6 +826,19 @@ function time(data){
 	return final;
 }
 
+function toString(num){
+	return num + "";
+}
+
+function toInt(string){
+	return parseInt(string);
+}
+
+function wait(seconds,callback){
+	var ticks = seconds * 1000;
+	setTimeout(function(){callback();}, ticks);
+}
+
 function loop(times,callback){
 	times -= 1;
 	for(var i = 0; i <= times; i++){
@@ -678,7 +865,6 @@ function onPressKey(idclass,letter,callback) {
 		}
 	};
 	}
-		
 	
 }